fix: 거리 정보(distanceMeters) 소수점 제거
- Haversine 계산 SQL 결과를 ROUND로 감싸 소스에서부터 정수 미터로 계산
- distanceMeters 필드를 Double -> Long으로 변경(MerchantResponseDto,
  NearbyMerchantRecommendationResponseDto, NearbyBenefitResponseDto,
  TodayCardRecommendationResponseDto.NearbyMerchant) - 응답이 "235.6789"나
  "235.0"이 아니라 "235"로 내려가도록
- 위 필드를 쓰던 정렬 비교자(Facade/BenefitServiceImpl/RecommendationServiceImpl)와
  관련 테스트를 Long 타입에 맞게 수정

+++ src/main/java/site/benepay/domain/recommendation/dto/NearbyMerchantRecommendationResponseDto.java
@@ -21,7 +21,7 @@ public class NearbyMerchantRecommendationResponseDto {
 	private String address;
 	private BigDecimal latitude;
 	private BigDecimal longitude;
-	private Double distanceMeters;
+	private Long distanceMeters;
 	private String phone;
 	// 지금 당장(즉시할인) 혜택을 쓸 수 있는 매장인지 - true일 때만 recommendedCards가 채워진다.
 	private boolean benefitAvailable;
